requirements.txt
google-cloud-speech>=2.26.0
python-dotenv>=1.0.0
sounddevice>=0.5.0
scipy>=1.11.0
requests>=2.28.0
numpy>=1.24.0
